backend/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.2
django-extensions==3.2.3
django-ninja==1.1.0
pydantic==2.6.1
pydantic_core==2.16.2
sqlparse==0.4.4
typing_extensions==4.9.0
